fix: 결제 금액 타입 Long 통일 및 테스트 toss 설정 추가

+++ src/main/java/org/example/groommvp/domain/payment/client/TossPaymentClient.java
@@ -25,7 +25,7 @@ public class TossPaymentClient {
 		this.restClient = RestClient.create();
 	}
 
-	public void confirm(String paymentKey, String orderId, int amount) {
+	public void confirm(String paymentKey, String orderId, Long amount) {
 		String encodedAuth = Base64.getEncoder()
 			.encodeToString((secretKey + ":").getBytes(StandardCharsets.UTF_8));
